# openapi.yaml
openapi: 3.0.3
info:
  title: QR API
  description: >
    Stateless QR code generation API. Returns an SVG image or the raw boolean
    module matrix. No data is persisted or logged.
  version: "1.0.0"
  contact:
    name: API support
    email: jonashaemmerle0504@gmail.com
servers:
  - url: https://qr-api.jay-trading.workers.dev
    description: Production

paths:
  /health:
    get:
      operationId: health
      summary: Liveness check
      security: []
      responses:
        "200":
          description: Service is up
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: ok }
                  timestamp: { type: string, format: date-time }

  /v1/qr:
    get:
      operationId: generateQrSvg
      summary: Generate a QR code as an SVG image
      parameters:
        - name: data
          in: query
          required: true
          schema: { type: string, maxLength: 2000 }
          description: Content to encode. Capped at 2000 characters.
        - name: ecc
          in: query
          schema: { type: string, enum: [L, M, Q, H], default: M }
          description: Error correction level — higher survives more damage/obstruction at the cost of a denser code.
        - name: cellSize
          in: query
          schema: { type: integer, minimum: 1, maximum: 50, default: 8 }
        - name: color
          in: query
          schema: { type: string, example: "000000" }
          description: Foreground (module) color, 3 or 6 digit hex, no leading '#'.
        - name: bgColor
          in: query
          schema: { type: string, example: "ffffff" }
          description: Background color, 3 or 6 digit hex, no leading '#'.
      responses:
        "200":
          description: QR code image
          content:
            image/svg+xml:
              schema: { type: string }
        "400": { $ref: "#/components/responses/BadRequest" }

  /v1/qr/matrix:
    get:
      operationId: generateQrMatrix
      summary: Generate a QR code as a raw boolean module matrix
      parameters:
        - name: data
          in: query
          required: true
          schema: { type: string, maxLength: 2000 }
        - name: ecc
          in: query
          schema: { type: string, enum: [L, M, Q, H], default: M }
      responses:
        "200":
          description: QR module matrix
          content:
            application/json:
              schema:
                type: object
                properties:
                  size: { type: integer, description: "Matrix is size x size modules" }
                  ecc: { type: string, enum: [L, M, Q, H] }
                  matrix:
                    type: array
                    items:
                      type: array
                      items: { type: boolean }
        "400": { $ref: "#/components/responses/BadRequest" }

components:
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }
              message: { type: string }
  securitySchemes:
    RapidAPIProxySecret:
      type: apiKey
      in: header
      name: X-RapidAPI-Proxy-Secret
security:
  - RapidAPIProxySecret: []